BlogProject/requirements.txt
alembic==1.7.4
amqp==5.0.6
aniso8601==9.0.1
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
attrs==21.4.0
billiard==3.6.4.0
Brlapi==0.6.6
cached-property==1.5.2
celery==5.1.2
certifi==2018.1.18
cffi==1.15.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.3.7
cryptography==3.3.2
cupshelpers==1.0
dataclasses==0.8
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.45.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-passwordreset==1.2.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
environs==9.3.4
filelock==3.4.1
flake8==4.0.0
Flask==2.0.2
Flask-Migrate==3.1.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.9.2
idna==2.6
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
JsonForm==0.0.2
jsonschema==4.0.0
JsonSir==0.0.2
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
mccabe==0.6.1
migrate==0.3.8
oauth==1.0.1
oauthlib==3.1.1
olefile==0.45.1
pbr==5.8.0
pexpect==4.2.1
phonenumbers==8.12.38
phonenumberslite==8.12.38
Pillow==5.1.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
protobuf==3.0.0
psycopg2==2.7.7
psycopg2-binary==2.9.2
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.4.0
PyGObject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyrsistent==0.18.0
python-apt==1.6.5+ubuntu0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.19.1
Python-EasyConfig==0.1.7
python3-openid==3.2.0
pytz==2021.3
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
Resource==0.2.1
SecretStorage==2.3.1
semantic-version==2.8.5
setuptools-rust==0.12.1
simplejson==3.13.2
six==1.16.0
SQLAlchemy==1.4.26
sqlparse==0.4.2
stevedore==3.5.0
system-service==0.3
systemd-python==234
toml==0.10.2
typing_extensions==4.0.0
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
vine==5.0.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.2
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
